алгебраическая форма
